Compatible with corrected PINK transform output
The initial hack to the PINK algorithm to output the transform information was not correct. I was only saving the first element of the bestRotationMatrix, which has the same shape as the euclidean distance heat map.

I have rewritten PINK to handle the correct, full output, and had to adjust the plotting facility to deal with it.