Initial comit of the RGZ FITS loading
This is the initial commit of the code that will load in the FITS data following the RGZ_RCNN objects, with the actual FITS downloading and reprojecting being carried out outside of the package.

This commit will load the FITS images, convert them to numpy arrays one disk, calidate the numpy arrays as they are being loaded and dump them to the binary dumpy file when request.

I will be removing thedownloading code in the next commit.