Initial commit of the log10 preprocessing function
The initial commit of the log10 preprocessing function. It will log the data and attempt to remove invalid pixels (i.e. NaNs of negative or zeros). It should
be done after the sigma clipping I think. As that will get rid of pixels below the noise and zero them. Once the logging is done than the normalisation should be applied.