Added Color plots, color transform scale, test expriments and oversampling
Created code to plots counts as a color for teh label attributes.

Added color mappings to the figures, and have applied a SymLogNorm transform to the colors to better distinguish the features in the dumped SOM layers

Added additional plotting modes when showing teh trained SOM

Additional test experiemtns have been added

Running pink with a factor of ten for the versampling when training the SOM. More iterations requires more time.