Initial commit of train/validate splitting
Added code to perform the initial steps of spliting the data into a training and validation data set

Added a wrapper around the dump_binary function, after it had been renamed to _write_binary(), to handle splitting the valid sources into two groups. This will be expanded after to be able to handle the cross-validation technique

Removed references to Dask. No longer being used to download or speed up operations.