Working commit of the fraction keyword
A working implementation of the fraction kewyord that is supplied to dump_binary(). This will split the sources into a `train` and `validation` set before dumping them out to disk.

Subsequent tasks, like map or count_map, also have a mode keyword to set which binary file to use when mapping or plotting