dded trials option
Add the trials option when producing the book object. It attempts to use the Euclidean distance as a proxy for likelihood. It will then draw some user specific number of radon draws following this likelihood matrix when placing labels.

It seems to be working, but the attribute_label plots are not showing any improved clustering. If anything, worse. I thought I had this working yesterday, but I think I actually had the option around the wrong way, so trials were being run when it was meant to be a single, closest match.