Validator structure created, saving results to dataframe
Created a dict returned by the Pink.validator() method that can be collected and placed into a DataFrame (for example). The dict will also contain items related to that set of processing - i.e. preprocessing options and pink arguments.

Also tweak the validator() method slightly to perform a once over on the answers to flatten them, before comparing them to each item in the valid binary. Seems to be a lot faster