First complete commit of segments
I think I have gotten al the funcitonality working with these segemented binary files. I had to rearrange a far chuck of code to keep things agnostic as to plotting and the training binary.

I think that there may be a bug and the same train binary is used during the validator(0 function. Will investigate this shortly.